requirements.txt
streamlit
python-docx
requests
pandas
openpyxl